+++ frontend/libs/studio-pure-functions/src/ArrayUtilss/ArrayUtilss.test.ts
@@ -1,4 +1,4 @@
-import { ArrayUtils } from './ArrayUtils';
+import { ArrayUtils } from './ArrayUtilss';
 
 describe('ArrayUtils', () => {
   describe('removeDuplicates', () => {
