--- frontend/libs/studio-pure-functions/src/arrayUtils/index.ts
@@ -1,1 +1,0 @@
-export * from './ArrayUtils';
